// src/apis/models/FireBytes_Unified_ButtonGroups_Dtos_ButtonConfigs.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_ButtonGroups_Dtos_ButtonConfigs = {
    sortNum?: number | null;
    type?: string | null;
    className?: string | null;
    eventName?: string | null;
    icon?: string | null;
    shape?: string | null;
    size?: string | null;
    width?: number | null;
    title?: string | null;
    acl?: string | null;
    color?: string | null;
    bgColor?: string | null;
    position?: string | null;
    disable?: boolean | null;
};