// src/apis/models/FireBytes_Unified_ButtonGroups_Dtos_ButtonGroupsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_ButtonGroups_Dtos_ButtonConfigs } from './FireBytes_Unified_ButtonGroups_Dtos_ButtonConfigs';
export type FireBytes_Unified_ButtonGroups_Dtos_ButtonGroupsDto = {
    buttonConfigs?: Array<FireBytes_Unified_ButtonGroups_Dtos_ButtonConfigs> | null;
};